chore: Move v4 iOS code to legacy directory
