+++ ios/legacy/RCTImageComponentView+RNSScreenStackHeaderConfig.h
