+++ ios/legacy/RCTImageComponentView+RNSScreenStackHeaderConfig.mm
