+++ ios/legacy/RNSScreen.mm
@@ -19,7 +19,6 @@
 #import "RNSHeaderHeightChangeEvent.h"
 #import "RNSScreenViewEvent.h"
 
-#import "RNSConversions.h"
 #import "RNSSafeAreaViewComponentView.h"
 #import "RNSSafeAreaViewNotifications.h"
 #import "RNSScreenFooter.h"
@@ -28,6 +27,7 @@
 #import "RNSScrollViewFinder.h"
 #import "RNSScrollViewHelper.h"
 #import "RNSTabBarController.h"
+#import "conversion/RNSConversions.h"
 
 #import "RNSDefines.h"
 #import "UIView+RNSUtility.h"
