+++ ios/legacy/RNSScreenFooter.h
