+++ ios/legacy/RNSScreenFooter.mm
