+++ ios/legacy/RNSScreenStackAnimator.mm
