+++ ios/legacy/RNSScreenStackHeaderConfig.h
