+++ ios/legacy/RNSScreenStackHeaderConfig.mm
