+++ ios/legacy/RNSScreenWindowTraits.h
