+++ ios/legacy/RNSScreenWindowTraits.mm
