+++ ios/legacy/events/RNSHeaderHeightChangeEvent.mm
