+++ ios/legacy/integrations/RNSDismissibleModalProtocol.h
