+++ ios/legacy/utils/UINavigationBar+RNSUtility.h
